Move `TestAutomation` location.
